refactor: move force_install.rs tests to a sibling force_install_tests.rs
force_install.rs had ~140 lines of inline #[cfg(test)] mod tests,
violating .rules/rust.md #4 (non-trivial tests belong in a sibling
tests.rs file). Extracted verbatim into plugins/force_install_tests.rs,
following the same pattern as top-level modules/models_tests.rs,
paths_tests.rs, etc. — declared in plugins/mod.rs as
#[cfg(test)] mod force_install_tests;, importing the tested items via
super::force_install::{...} instead of super::*.

No behavioral change: all 10 tests moved as-is and still pass under
their new module path (plugins::force_install_tests::*).
