+++ src-tauri/src/plugins/mod.rs
@@ -3,6 +3,8 @@ pub mod compat; // COMPAT(registry-ga): remove with the BC layer
 pub mod deep_link;
 pub mod driver;
 pub mod force_install;
+#[cfg(test)]
+mod force_install_tests;
 pub mod install_cancellation;
 pub mod installer;
 pub mod integrity;
